\name{indbourse}
\alias{indbourse}
\docType{data}
\title{Stock price indices for the period 2006-2010}
\description{
indbourse contains stock price indices for the period 2006-2010 : Nikkei (name of the quote symbol: N225), ESTX50 EURP (STOXX50E), 
Dow Jones (DJI), Nasdaq (IXIC), CAC40 (^FCHI) and PARIS IND SBF120 (SBF120).  
}
\format{indbourse is an \code{its} object.}
\source{Yahoo finance}
\examples{
data(indbourse)
}
\keyword{datasets}
