\name{m30}
\alias{m30}
\docType{data}
\title{Fatalities in car accidents in France for the period 1973-2006}
\description{
\code{m30} is the series of monthly fatalities in car accidents in France for the period 1973-2006.}
\usage{data(m30)}
\format{Time series data}
\details{The data from July 1973 to December 2004 have been multiplied by 1.069 to take 
into account the change of the definition of a fatal accident. 
Until 2004, an accident is fatal if death occurs within 6 days whereas from 
2006 the deadline moves to 30 days.
}
\source{http://www.securite-routiere.org/}
\examples{
data(m30)
}
\keyword{datasets}
