\name{popfr}
\alias{popfr}
\docType{data}
\title{French population for the period 1846-1951}
\description{
\code{popfr} is the French population average for the period 1846-1951.}
\usage{data(popfr)}
\format{Time series data}
\details{One unit: 1000 inhabitants. Two missing values in 1916 and 1941}
\source{https://www.insee.fr/fr/statistiques}
\examples{
data(popfr)
}
\keyword{datasets}
