\name{caschrono-package}
\alias{caschrono-package}
\alias{caschrono}
\docType{package}
\title{
Series Temporelles Avec R
}
\description{
Functions, data sets and exercises solutions for the book 'Séries temporelles avec R' (Yves Aragon,
edp sciences, 2016). For all chapters, a vignette is available with some additional material and exercises solutions.  
}
\details{
\tabular{ll}{
Package: \tab caschrono\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2016-08-16\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Yves Aragon,
Maintainer: Thibault Laurent <thibault.laurent@univ-tlse1.fr>
}
\references{
Y. Aragon (2016), Séries temporelles avec R, edp sciences
}

\keyword{ts}
