\name{caschrono-package}
\alias{caschrono-package}
\alias{caschrono}
\docType{package}
\title{
Series Temporelles avec R - Methodes et cas (Time Series with R: methods and applications) 
}
\description{
Functions, data sets and exercises solutions for the book 'Séries temporelles avec R - Méthodes et cas' (Yves Aragon,
Springer, 2011). For all chapters, a vignette is available with some additional material and exercises solutions.  
}
\details{
\tabular{ll}{
Package: \tab caschrono\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-21\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Yves Aragon,
Maintainer: Thibault Laurent <thibault.laurent@univ-tlse1.fr>
}
\references{
Y. Aragon (2011), Séries temporelles avec R - Méthodes et cas, Springer
}

\keyword{ts}
