\name{lait}
\alias{lait}
\docType{data}
\title{Milk collection in France}
\description{\code{lait} is the monthly  milk collection in France,  January 1980 - January 2010}
\format{\code{lait} is a \code{ts} object}
\details{Data are expressed in thousands of tons}
\source{Enquête laitière mensuelle - Service de la Statistique et de la Prospective (SSP) - Ministère de l'Alimentation, 
de l'Agriculture et de la Pêche.}
\examples{
data(lait)
}
\keyword{datasets}
