\name{popfr}
\alias{popfr}
\docType{data}
\title{French population for the period 1846-1951}
\description{
\code{popfr} is the French population average for the period 1846-1951.}
\usage{data(popfr)}
\format{Time series data}
\details{One unit: 1,000 inhabitants. Two missing values in 1916 and 1941}
\source{http://www.insee.fr/fr/themes/tableau.asp?ref_id=NATnon02145     \cr
http://dev-www.ined.fr/fr/pop_chiffres/france/structure_population/recensement_1846/}
\examples{
data(popfr)
}
\keyword{datasets}
